Keep Vulkan policy tests with their owner
